# cmd/api/docs/swagger.yaml
basePath: /api
definitions:
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        type: string
      email:
        type: string
      expires_at:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      refresh_token:
        type: string
      user_id:
        type: string
    type: object
  dto.LogoutOtherRequest:
    properties:
      current_refresh_token:
        type: string
    required:
    - current_refresh_token
    type: object
  dto.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RefreshTokenResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      refresh_token:
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        description: Email is now optional
        type: string
      first_name:
        type: string
      last_name:
        type: string
      otp_code:
        description: OTP code is required
        type: string
      password:
        minLength: 8
        type: string
      phone_number:
        description: Phone number is now required
        type: string
    required:
    - first_name
    - last_name
    - otp_code
    - password
    - phone_number
    type: object
  dto.RegisterResponse:
    properties:
      access_token:
        type: string
      email:
        type: string
      expires_at:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      refresh_token:
        type: string
      user_id:
        type: string
    type: object
  dto.RegisterWithOTPRequest:
    properties:
      code:
        description: OTP code to verify
        type: string
      email:
        description: Email is optional
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      phone_number:
        type: string
    required:
    - code
    - first_name
    - last_name
    - password
    - phone_number
    type: object
  dto.RequestOTPRequest:
    properties:
      phone_number:
        type: string
    required:
    - phone_number
    type: object
  dto.RequestOTPResponse:
    properties:
      expires_in:
        description: seconds until expiry
        type: integer
      message:
        type: string
    type: object
...
  dto.VerifyOTPRequest:
    properties:
      code:
        type: string
      phone_number:
        type: string
    required:
    - code
    - phone_number
    type: object
  dto.VerifyOTPResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      is_new_user:
        type: boolean
      phone_number:
        type: string
      refresh_token:
        type: string
      user_id:
        type: string
    type: object
  errors.ErrorResponse:
    properties:
      code:
        type: string
      errors:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@auth-guardian.io
    name: API Support
    url: http://www.auth-guardian.io/support
  description: Authentication and authorization service with advanced security features.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Auth Guardian API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and receive access and refresh tokens
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the current user session
      parameters:
      - description: Logout request with refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/logout-all:
    post:
      consumes:
      - application/json
      description: Invalidate all active sessions for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout from all devices
      tags:
      - auth
  /auth/logout-others:
    post:
      consumes:
      - application/json
      description: Invalidate all sessions except the current one
      parameters:
      - description: Current session refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutOtherRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout from other devices
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a valid refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/register-with-otp:
    post:
      consumes:
      - application/json
      description: Register a new user with OTP verification
      parameters:
      - description: Registration data with OTP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterWithOTPRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Invalid OTP
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Register with OTP
      tags:
      - auth
  /auth/request-otp:
    post:
      consumes:
      - application/json
      description: Request a one-time password for authentication
      parameters:
      - description: Email address for OTP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RequestOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RequestOTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Request OTP
      tags:
      - auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Verify a one-time password
      parameters:
      - description: OTP verification data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VerifyOTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Invalid OTP
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Verify OTP
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'Bearer Token Authentication. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"